+++ src/frontend/js/types/User.ts
@@ -1,0 +1,4 @@
+export interface User {
+  full_name: string;
+  username: string;
+}